Create Eye Controlled Mouse
The eye camera follows the head movements maintaining the pupil-centered image. When a tracking error
is produced, the image from a camera with a wider field of view is used to locate the eye and quickly
recover the tracking process. Its special shape has been exploited to allow the optimization of the image
processing algorithms developed for this system.